+++ src/wp/ap/xp/ap_Menu_ActionSet.cpp
@@ -467,6 +467,7 @@ EV_Menu_ActionSet * AP_CreateMenuActionSet(void)
 	_s(AP_MENU_ID_HELP_SEARCH,		0,0,0,0,	"helpSearch",			nullptr,					ap_GetLabel_Search);
 	_s(AP_MENU_ID_HELP_ABOUT,		0,1,0,0,	"dlgAbout",			nullptr,					ap_GetLabel_About);
 	_s(AP_MENU_ID_HELP_REPORT_BUG, 0,0,0,0, "helpReportBug", nullptr, nullptr);
+	_s(AP_MENU_ID_HELP_CHANGELOG,	0,0,0,0,	"helpChangelog",		nullptr,					nullptr);
 
 #ifdef ENABLE_SPELL
 	_s(AP_MENU_ID_SPELL_SUGGEST_1,	0,0,0,0,	"spellSuggest_1",	ap_GetState_Suggest,	ap_GetLabel_Suggest);