+++ src/wp/ap/xp/ap_Menu_Id_List.h
@@ -425,6 +425,7 @@ menuitem(HELP_CHECKVER)
 menuitem(HELP_SEARCH)
 menuitem(HELP_ABOUT)
 menuitem(HELP_REPORT_BUG)
+menuitem(HELP_CHANGELOG)
 
 menuitem(SPELL_SUGGEST_1)		// _suggest_1 thru _suggest_9 must be contiguous
 menuitem(SPELL_SUGGEST_2)